requirements.txt
# Brownie Framework
eth-brownie>=1.19.0

# Vyper Compiler
vyper>=0.3.10,<0.4.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Static Analysis
slither-analyzer>=0.9.0

# Environment Variables
python-dotenv>=1.0.0

